the app modules rename package name
