+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion rootProject.ext.compile_sdk_version
 
     defaultConfig {
-        applicationId "com.dvsnier.demo.toolchain"
+        applicationId "com.dvsnier.toolchain.demo"
         minSdkVersion rootProject.ext.min_sdk_version
         targetSdkVersion rootProject.ext.target_sdk_version
         versionCode rootProject.ext.app_version_code
