+++ app/src/main/AndroidManifest.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.dvsnier.demo">
+    package="com.dvsnier.toolchain.demo">
 
     <application
-        android:name=".DvsnierApplication"
+        android:name=".application.DvsnierApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
